safari-600.7-branch/LayoutTests/js/dfg-strict-mode-to-this-expected.txt

Tests that doing to-this conversion in strict mode doesn't cause us to believe that if the input is an object then the output is also an object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS thingy() is false on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
